Recipes now have UIDs.
Because recipes use to be found by name only, then recipes with the same name displayed the first added item (fixed by passing in the list of recipes to the Adapter rather than their names) and deleting any of them deleted the first recipe. Giving recipes a UID fixes both of these issues.