Add an EditText right below the one that receives the Enter key
While the final implementation for this is quite simple, it took me quite a while to find it! I started with just adding a onKey() to every EditText, but that introduced a cyclic dependency because I would add a new EditText there which needed its own onKey, and so on. Then, I tried subclassing EditText, so I could implement onKey() there, but I couldn't get it to work with the XML. Finally, I doubled back and found the callback when a child view is added and everything went smoothly from there.